requirements.txt
llama-index-core
llama-index-llms-ollama
pandas
numpy
matplotlib
scipy